// src/types/types.ts
/** @format */

export interface listOfWeatherObject {
	humidity: number;
	temp: number;
	temp_max: number;
	temp_min: number;
	speed: number;
	pressure: number;
	description: number;
	date: string;
	id: number;
}
[];

export interface detailOfLocation {
	name: string;
	country: string;
}

export interface weatherobject {
	humidity: number;
	temp: number;
	temp_max: number;
	temp_min: number;
	speed: number;
	pressure: number;
	description: number;
	date: string;
	id: number;
}

export interface locInfo {
	name: string;
	country: string;
}